* Fire   * New: Added a new drum mode, see manual for details.   * New: The last selected note/sequencer view is remembered when switching to session/mix and back.   * New: All LEDs are full turned off on shutdown and a Goodbye message is displayed.   * New: Session mode: ALT + pad selects clip (without starting playback).   * New: Bank button + Touch knob: Resets the knobs value to the default value.   * Fixed: Black pads were not fully turned off.   * Fixed: Do not leave Mix mode on track selection. * Komplete Kontrol MkI   * New: Works now on Apple Silicon.   * Fixed: USB resources were not freed on shutdown. * Launchkey   * New: Added detection for Launchkey 88 * MCU   * Fixed: Knob LEDs went off when panorama was panned fully left.

+++ src/main/java/de/mossgrabers/reaper/framework/daw/CursorClipImpl.java
@@ -242,7 +242,7 @@ public class CursorClipImpl extends BaseImpl implements INoteClip
 
     /** {@inheritDoc} */
     @Override
-    public void setLoopLength (final int length)
+    public void setLoopLength (final double length)
     {
         this.setPlayEnd (this.getPlayStart () + length);
     }
